keep mcp stuff together
